fix(lime-plugin-metrics): fix retry
