+++ plugins/lime-plugin-metrics/src/metricsApi.js
@@ -1,21 +1,5 @@
 import api from "utils/uhttpd.service";
 
-import {
-    setMetricsQueryData,
-    setMetricsQueryDataToFetching,
-} from "./metricsQueries";
-
-export const getAllMetrics = async (ips) => {
-    const metrics = [];
-    for (const ip of ips) {
-        setMetricsQueryDataToFetching(ip);
-        const metric = await getMetrics(ip);
-        metrics.push({ ip: metric });
-        setMetricsQueryData(ip, metric);
-    }
-    return metrics;
-};
-
 export const getMetrics = (ip) => {
     console.log("getting metrics for ", ip);
     return api.call("lime-metrics", "get_metrics", { target: ip });
