+++ plugins/lime-plugin-metrics/src/metricsQueries.js
@@ -2,33 +2,36 @@ import { useQuery } from "@tanstack/react-query";
 
 import queryCache from "utils/queryCache";
 
-import { getAllMetrics, getGateway, getMetrics, getPath } from "./metricsApi";
-
-/**
- * Util function to update query data of a single metrics ip using queryCache
- * Is used to set the {status: 'fetching'} or the metric itself from anywhere
- */
-export function setMetricsQueryData(ip, data) {
-    queryCache.setQueryData(["lime-metrics", "get_metrics", ip], data);
-}
-
-export function setMetricsQueryDataToFetching(ip) {
-    setMetricsQueryData(ip, { status: "fetching" });
-}
+import { getGateway, getMetrics, getPath } from "./metricsApi";
 
 export function useMetrics(ip, params) {
-    return useQuery(
-        ["lime-metrics", "get_metrics", ip],
-        () => getMetrics(ip),
-        params
-    );
+    return useQuery(["lime-metrics", "get_metrics", ip], () => getMetrics(ip), {
+        retry: false,
+        ...params,
+    });
 }
+
+export const getAllMetrics = async (ips) => {
+    const metrics = [];
+    for (const ip of ips) {
+        const metric = await queryCache.fetchQuery([
+            "lime-metrics",
+            "get_metrics",
+            ip,
+        ]);
+        metrics.push({ ip: metric });
+    }
+    return metrics;
+};
 
 export function useAllMetrics(ips, params) {
     return useQuery(
         ["lime-metrics", "get_metrics", ips],
         (query) => getAllMetrics(query.queryKey[2]),
-        params
+        {
+            retry: false,
+            ...params,
+        }
     );
 }
